backend/test_requirements.txt
pytest>=6.1.0
coverage>=5.3
pytest-cov>=2.10.1
requests-mock>=1.8.0
mock>=4.0.2